+++ packages/common/web/src/Utils/draftUtils.js
@@ -83,6 +83,30 @@ function addEntity(editorState, targetSelection, entityData) {
   return EditorState.push(editorState, newContentState, 'apply-entity');
 }
 
+export const hasLinksInBlock = (block, contentState) => {
+  try {
+    //editor
+    if (block.findEntityRanges) {
+      return !!getLinkRangesInBlock(block, contentState).length;
+    }
+    //viewer
+    if (block.entityRanges && block.entityRanges.length) {
+      return block.entityRanges.some(entityRange => {
+        const entityMap = contentState.get('entityMap');
+        const entityKey = entityMap[entityRange.key];
+        const entity = contentState.getEntity(entityKey);
+        const entityType = entity.type;
+        if (entityType === 'LINK' || entityType === 'wix-draft-plugin-external-link') {
+          return true;
+        }
+        return false;
+      });
+    }
+  } catch (e) {
+    return false;
+  }
+};
+
 export const hasLinksInSelection = editorState => {
   return !!getSelectedLinks(editorState).length;
 };