+++ packages/plugin-hashtag/web/src/HashtagDecorator.jsx
@@ -1,5 +1,6 @@
 import Immutable from 'immutable';
 import { range } from 'lodash';
+import { hasLinksInBlock } from 'wix-rich-content-common';
 import Hashtag from './HashtagComponent';
 import hashtagRegexes from './hashtagRegexes';
 
@@ -8,13 +9,18 @@ export default class HashtagDecorator {
     this.componentProps = componentProps;
   }
 
-  getDecorations(block) {
+  getDecorations(block, contentState) {
     const key = block.getKey();
     const text = block.getText();
     const type = block.getType();
     const decorations = Array(text.length).fill(null);
 
-    if (type === 'code-block' || !text || !text.match(hashtagRegexes.hashSigns)) {
+    if (
+      type === 'code-block' ||
+      !text ||
+      hasLinksInBlock(block, contentState) ||
+      !text.match(hashtagRegexes.hashSigns)
+    ) {
       return Immutable.List(decorations); // eslint-disable-line new-cap
     }