+++ CHANGELOG.md
@@ -18,7 +18,9 @@
   - [#380](https://github.com/wix-incubator/rich-content/pull/380) Image Editor (Wix Image Studio)
 ### :bug: Bug Fix
   - `html`
-    - [#385](  https://github.com/wix-incubator/rich-content/pull/385) edit panel resets to default state when focusing on different instances of the same plugin
+    - [#385](https://github.com/wix-incubator/rich-content/pull/385) edit panel resets to default state when focusing on different instances of the same plugin
+  - `viewer`
+  - [#386](https://github.com/wix-incubator/rich-content/pull/386) code block duplication
 <hr/>
 
 ## 5.0.1 (Sep 10, 2019)